requirements.txt
pyyaml
pydantic
typer
rich
sentence-transformers
faiss-cpu
numpy